move version fetching to shared folder
